Restructured the program, logging and json config
Lay out the structure of the program by creating functions for the main
logic and functionality of finding links for downloading from the site.
Added in logging capability - both filehandler and stream handler.
Added in loading of magazine download configurations from a default
configuration file. If this file doesn't exist, the program will create
one with some default magazines to download.